Move ipbus parser logic to separate module
