+++ ipbus_parser/__init__.py
@@ -1,0 +1,3 @@
+from .packet import Packet
+
+__all__ = ["Packet"]
