+++ ipbus_parser/parameters.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class Endianness(Enum):
+    BIG = "big"
+    LITTLE = "little"
